fix: strip null bytes when parsing the vega20 format and add tests with real 6700xt data
For some reason the clocks tables on rx 6000 series GPUs include null
bytes on some of the lines. The vega20 parser has been updated to trim null
bytes on all lines, and a test was added with `pp_od_clk_voltage` being
taken directly from the SysFS (as opposed to being copied as text).