feat: add basic support for RDNA3 table format, rework ranges to use signed integers
Adjust the parser to work on the table provided in https://github.com/ilya-zlobintsev/LACT/issues/190#issuecomment-1660846115.
This required some changes in the table representation, in particular all ranges are now signed integers.
The per-power-level clockspeed and voltage ranges in OD_RANGE are now
decoupled, as RDNA3 only seems to provide voltage there.

Bumps version to 0.12.0.